requirements.txt
contourpy==1.1.0
cycler==0.11.0
fonttools==4.42.0
imageio==2.31.1
kiwisolver==1.4.4
lazy_loader==0.3
matplotlib==3.7.2
networkx==3.1
numpy==1.25.2
packaging==23.1
pandas==2.1.1
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
relativeNeighborhoodGraph==0.0.1
scikit-image==0.21.0
scipy==1.11.1
shapely==2.0.1
six==1.16.0
tifffile==2023.7.18
scikit-learn==1.3.0
./GDAL-3.4.3-cp311-cp311-win_amd64.whl
